requirements.txt
flask
mysql-python
flask-sqlalchemy
flask-script
flask-wtf
flask-migrate
py-bcrypt
python-slugify
flask-markdown
flask-uploads
Flask-Social
http://github.com/pythonforfacebook/facebook-sdk/tarball/master
python-twitter
foursquare
google-api-python-client
Flask-OAuth
Flask-Login==0.3.1
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.9.6
argparse==1.2.1
itsdangerous==0.24
rauth==0.7.0
requests==2.4.3
Flask-Mail
flask-sendgrid